flow-based codezoo/SfmSR-main/requirements.txt
accelerate==0.33.0
diffusers==0.32.1
huggingface_hub==0.23.4
lpips==0.1.4
matplotlib==3.7.5
numpy==1.24.4
openai_clip==1.0.1
opencv_python==4.10.0.84
packaging==24.2
pandas==2.0.3
Pillow==11.1.0
scipy==1.15.1
skimage
thop
torch==2.4.0+cu124
torchvision==0.19.0+cu124
tqdm==4.66.4
wandb==0.19.4

